move console output to new file "console.js"
