+++ lib/engine.js
@@ -32,7 +32,8 @@ module.exports.start = function start(configuration) {
       configuration.emitter.on(listener.type, listener.listener);
     });
   }
-
+  // attach console listener
+  configuration.emitter.on('master:end', require('./console'));
   let _flow = flow.map(function (fn) {
     return fn.bind(configuration);
   });
