+++ lib/instance.js
@@ -110,16 +110,16 @@ module.exports = function instance(input, instanceComplete, progress) {
     runner.on('suite', function (Suite) {
       log('suite event, suite %s, root: %s', Suite.title, Suite.root);
       if (!(nemo && nemo.driver)) {
-        Suite.beforeAll(function checkNemo(_doneMochaBeforeAll) {
+        Suite.beforeAll(function checkNemo(suiteBeforeComplete) {
           Nemo(config).then(function (_nemo) {
             nemo = _nemo;
             nemo.runner = {
               context: input
             };
-            _doneMochaBeforeAll();
+            suiteBeforeComplete();
           }).catch(function (err) {
             console.error(err);
-            _doneMochaBeforeAll(err);
+            suiteBeforeComplete(err);
             instanceComplete(err);
           });
         });
