+++ lib/master.js
@@ -39,10 +39,13 @@ module.exports = function master() {
     iconfig = this.instances[0];
 
     log('single kickoff with instance %O', iconfig.tags);
-    // TODO: below doesn't impact debuglogs of modules already loaded
     process.env = Object.assign({}, process.env, iconfig.conf.env || {});
     iconfig.tags.uid = uuid();
-    return instance({basedir: this.program.baseDirectory, profile: iconfig, __dirname: __dirname}, err => console.error(err), progress);
+    return instance({basedir: this.program.baseDirectory, profile: iconfig, __dirname: __dirname}, err => {
+      if (err) {
+        console.error(err);
+      }
+    }, progress);
   }
   // parallel use case
   tasks = this.instances.map(function (iconf) {
